// src/app/generated_services/model/dayOfWeek.ts
/**
 * CarlsonGracie Backend API
 * API for managing students, belts, graduation requirements, notices, frequencies and user authentication with Keycloak
 *
 * Contact: marcosturossi@gmail.com
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


export type DayOfWeek = 'Sunday' | 'Monday' | 'Tuesday' | 'Wednesday' | 'Thursday' | 'Friday' | 'Saturday';

export const DayOfWeek = {
    Sunday: 'Sunday' as DayOfWeek,
    Monday: 'Monday' as DayOfWeek,
    Tuesday: 'Tuesday' as DayOfWeek,
    Wednesday: 'Wednesday' as DayOfWeek,
    Thursday: 'Thursday' as DayOfWeek,
    Friday: 'Friday' as DayOfWeek,
    Saturday: 'Saturday' as DayOfWeek
};